+++ app/public/wp-content/plugins/homburg-dealerportaal/includes/class-hdp-blocks.php
@@ -23,6 +23,7 @@ class HDP_Blocks {
 		register_block_type( HDP_PLUGIN_DIR . 'blocks/downloads' );
 		register_block_type( HDP_PLUGIN_DIR . 'blocks/content' );
 		register_block_type( HDP_PLUGIN_DIR . 'blocks/info-kaart' );
+		register_block_type( HDP_PLUGIN_DIR . 'blocks/contact-kaart' );
 		register_block_type( HDP_PLUGIN_DIR . 'blocks/portaal-kaart' );
 		register_block_type( HDP_PLUGIN_DIR . 'blocks/bestelgeschiedenis' );
 	}
@@ -34,6 +35,7 @@ class HDP_Blocks {
 				&& ! has_block( 'homburg/downloads-pagina', $post )
 				&& ! has_block( 'homburg/content-pagina', $post )
 				&& ! has_block( 'homburg/info-kaart', $post )
+				&& ! has_block( 'homburg/contact-kaart', $post )
 				&& ! has_block( 'homburg/portaal-kaart', $post )
 				&& ! has_block( 'homburg/bestelgeschiedenis-pagina', $post ) )
 		) {